Extract createProgram from applyTypes to allow sharing with instrument
