+++ src/compiler-helper.spec.ts
@@ -1,9 +1,9 @@
 import * as ts from 'typescript';
-import { applyTypes } from './apply-types';
+import { getProgram } from './compiler-helper';
 
-describe('applyTypes', () => {
+describe('getProgram', () => {
     it('should throw an error if given non-existing tsconfig.json file', () => {
-        expect(() => applyTypes([], { tsConfig: 'not-found-file.json' })).toThrowError(
+        expect(() => getProgram({ tsConfig: 'not-found-file.json' })).toThrowError(
             `Error while reading not-found-file.json: The specified path does not exist: 'not-found-file.json'.`,
         );
     });
@@ -13,7 +13,7 @@ describe('applyTypes', () => {
             ...ts.sys,
             readFile: jest.fn(() => '<invalid json>'),
         };
-        expect(() => applyTypes([], { tsConfig: 'tsconfig.bad.json', tsConfigHost })).toThrowError(
+        expect(() => getProgram({ tsConfig: 'tsconfig.bad.json', tsConfigHost })).toThrowError(
             `Error while reading tsconfig.bad.json: '{' expected.`,
         );
         expect(tsConfigHost.readFile).toHaveBeenCalledWith('tsconfig.bad.json');
@@ -24,7 +24,7 @@ describe('applyTypes', () => {
             ...ts.sys,
             readFile: jest.fn(() => '{ "include": 123 }'),
         };
-        expect(() => applyTypes([], { tsConfig: 'tsconfig.invalid.json', tsConfigHost })).toThrowError(
+        expect(() => getProgram({ tsConfig: 'tsconfig.invalid.json', tsConfigHost })).toThrowError(
             `Error while parsing tsconfig.invalid.json: Compiler option 'include' requires a value of type Array.`,
         );
         expect(tsConfigHost.readFile).toHaveBeenCalledWith('tsconfig.invalid.json');
